fix: problem comparing subcategories fix: use Int instead of BigInteger

+++ src/main/kotlin/org/soneira/savings/api/dto/CategoryDTO.kt
@@ -1,5 +1,3 @@
 package org.soneira.savings.api.dto
 
-import java.math.BigInteger
-
-data class CategoryDTO(val id: BigInteger, val description: String)
+data class CategoryDTO(val id: Int, val description: String)
