+++ src/main/kotlin/org/soneira/savings/infrastructure/persistence/mongo/document/ExpenseDocument.kt
@@ -2,13 +2,12 @@ package org.soneira.savings.infrastructure.persistence.mongo.document
 
 import org.springframework.data.mongodb.core.mapping.Field
 import org.springframework.data.mongodb.core.mapping.FieldType.DECIMAL128
-import org.springframework.data.mongodb.core.mapping.FieldType.INT64
+import org.springframework.data.mongodb.core.mapping.FieldType.INT32
 import java.math.BigDecimal
-import java.math.BigInteger
 
 class ExpenseDocument(
-    @Field(targetType = INT64)
-    val key: BigInteger,
+    @Field(targetType = INT32)
+    val key: Int,
     @Field(targetType = DECIMAL128)
     val value: BigDecimal
 )
