+++ src/main/kotlin/org/soneira/savings/infrastructure/repository/CategoryRepositoryImpl.kt
@@ -5,7 +5,6 @@ import org.soneira.savings.domain.repository.CategoryRepository
 import org.soneira.savings.infrastructure.persistence.mongo.mapper.CategoryMapper
 import org.soneira.savings.infrastructure.persistence.mongo.repository.CategoryMongoRepository
 import org.springframework.stereotype.Component
-import java.math.BigInteger
 
 @Component
 class CategoryRepositoryImpl(
@@ -17,5 +16,5 @@ class CategoryRepositoryImpl(
         return categories.map { categoryMapper.asCategory(it) }
     }
 
-    override fun getById(id: BigInteger) = getAll().first { it.id.value == id }
+    override fun getById(id: Int) = getAll().first { it.id.value == id }
 }
